Refactor to new start and utils modules.
